Move StopTimesHelper to its own package
