+++ src/main/java/org/opentripplanner/routing/RoutingService.java
@@ -18,6 +18,7 @@ import org.opentripplanner.routing.api.response.RoutingResponse;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.GraphIndex;
 import org.opentripplanner.routing.graphfinder.GraphFinder;
+import org.opentripplanner.routing.stoptimes.StopTimesHelper;
 import org.opentripplanner.standalone.server.Router;
 
 /**
